# openapi.yml
openapi: 3.1.0
info:
  title: MemoryOS browser API
  description: The same-origin contract consumed by the MemoryOS browser application.
  version: 1.0.0
servers:
- url: /
  description: Same-origin MemoryOS runtime
paths:
...
  /api/chat/sessions/{sessionId}/title:
    put:
      tags:
      - Chat
      summary: Rename an owned conversation
      operationId: renameChatSession
      parameters:
      - name: X-MemoryOS-CSRF
        in: header
        description: Same-origin non-simple request guard for browser-session mutations.
        required: true
        schema:
          type: string
          enum:
          - "1"
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Title"
        required: true
      responses:
        "400":
          description: Invalid chat request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "403":
          description: Tenant membership or CSRF requirement not met
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "404":
          description: Chat resource not accessible
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "409":
          description: Conversation is running or revision has changed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "401":
          description: Authentication required
        "200":
          description: Successful chat operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
      security:
      - bearerAuth: []
      - browserSession: []
    post:
      tags:
      - Chat
      summary: "Generate a short owned conversation title once, preserving manual\
        \ renames"
      operationId: generateChatTitle
      parameters:
      - name: X-MemoryOS-CSRF
        in: header
        description: Same-origin non-simple request guard for browser-session mutations.
        required: true
        schema:
          type: string
          enum:
          - "1"
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "400":
          description: Invalid chat request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "403":
          description: Tenant membership or CSRF requirement not met
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "404":
          description: Chat resource not accessible
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "409":
          description: Conversation is running or revision has changed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "401":
          description: Authentication required
        "200":
          description: "Current title, including fallback when naming is unavailable"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
      security:
      - bearerAuth: []
      - browserSession: []
...
  /api/chat/sessions/{sessionId}/settings:
    put:
      tags:
      - Chat
      summary: Atomically change assistant and project for subsequent turns
      operationId: configureChatSession
      parameters:
      - name: X-MemoryOS-CSRF
        in: header
        description: Same-origin non-simple request guard for browser-session mutations.
        required: true
        schema:
          type: string
          enum:
          - "1"
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSessionSettings"
        required: true
      responses:
        "400":
          description: Invalid chat request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "403":
          description: Tenant membership or CSRF requirement not met
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "404":
          description: Chat resource not accessible
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "409":
          description: Conversation is running or revision has changed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "401":
          description: Authentication required
        "200":
          description: Updated conversation settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
      security:
      - bearerAuth: []
      - browserSession: []
  /api/chat/sessions/{sessionId}/project:
    put:
      tags:
      - Chat
      summary: Move an owned conversation into or out of an owned project
      operationId: moveChatProject
      parameters:
      - name: X-MemoryOS-CSRF
        in: header
        description: Same-origin non-simple request guard for browser-session mutations.
        required: true
        schema:
          type: string
          enum:
          - "1"
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectSelection"
        required: true
      responses:
        "400":
          description: Invalid chat request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "403":
          description: Tenant membership or CSRF requirement not met
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "404":
          description: Chat resource not accessible
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "409":
          description: Conversation is running or revision has changed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "401":
          description: Authentication required
        "200":
          description: Successful chat operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
      security:
      - bearerAuth: []
      - browserSession: []
  /api/chat/sessions/{sessionId}/persona:
    put:
      tags:
      - Chat
      summary: Choose an authorized assistant for subsequent turns
      operationId: selectChatPersona
      parameters:
      - name: X-MemoryOS-CSRF
        in: header
        description: Same-origin non-simple request guard for browser-session mutations.
        required: true
        schema:
          type: string
          enum:
          - "1"
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonaSelection"
        required: true
      responses:
        "400":
          description: Invalid chat request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "403":
          description: Tenant membership or CSRF requirement not met
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "404":
          description: Chat resource not accessible
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "409":
          description: Conversation is running or revision has changed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "401":
          description: Authentication required
        "200":
          description: Successful chat operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
      security:
      - bearerAuth: []
      - browserSession: []
...
  /api/chat/sessions/{sessionId}/branch:
    put:
      tags:
      - Chat
      summary: Select a message version without generating a reply
      operationId: selectChatBranch
      parameters:
      - name: X-MemoryOS-CSRF
        in: header
        description: Same-origin non-simple request guard for browser-session mutations.
        required: true
        schema:
          type: string
          enum:
          - "1"
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchSelection"
        required: true
      responses:
        "400":
          description: Invalid chat request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "403":
          description: Tenant membership or CSRF requirement not met
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "404":
          description: Chat resource not accessible
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "409":
          description: Conversation is running or revision has changed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "401":
          description: Authentication required
        "204":
          description: No Content
      security:
      - bearerAuth: []
      - browserSession: []
...
  /api/chat/sessions/{sessionId}:
    get:
      tags:
      - Chat
      summary: Read an owned chat session
      operationId: getChatSession
      parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "400":
          description: Invalid request or cursor
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "403":
          description: Tenant membership or CSRF requirement not met
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "404":
          description: Conversation or message not accessible
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "401":
          description: Authentication required
        "200":
          description: Owned session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
      security:
      - bearerAuth: []
      - browserSession: []
    delete:
      tags:
      - Chat
      summary: Delete an owned conversation and stop its active reply
      operationId: deleteChatSession
      parameters:
      - name: X-MemoryOS-CSRF
        in: header
        description: Same-origin non-simple request guard for browser-session mutations.
        required: true
        schema:
          type: string
          enum:
          - "1"
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "400":
          description: Invalid chat request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "403":
          description: Tenant membership or CSRF requirement not met
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "404":
          description: Chat resource not accessible
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "409":
          description: Conversation is running or revision has changed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "401":
          description: Authentication required
        "204":
          description: No Content
      security:
      - bearerAuth: []
      - browserSession: []
...
  /api/chat/sessions/{sessionId}/branches:
    get:
      tags:
      - Chat
      summary: Read version relationships in an owned conversation
      operationId: getChatBranches
      parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "400":
          description: Invalid chat request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "403":
          description: Tenant membership or CSRF requirement not met
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "404":
          description: Chat resource not accessible
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "409":
          description: Conversation is running or revision has changed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "401":
          description: Authentication required
        "200":
          description: Successful chat operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatBranch"
      security:
      - bearerAuth: []
      - browserSession: []
...
components:
  schemas:
...
    Title:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 0
      required:
      - title
    ChatSession:
      type: object
      properties:
        id:
          type: string
          format: uuid
        personaId:
          type: string
          format: uuid
        rootMessageId:
          type: string
          format: uuid
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        projectId:
          type:
          - string
          - "null"
          format: uuid
      required:
      - createdAt
      - id
      - personaId
      - projectId
      - rootMessageId
      - title
      - updatedAt
...
    ChatSessionSettings:
      type: object
      properties:
        personaId:
          type: string
          format: uuid
        projectId:
          oneOf:
          - type: string
            format: uuid
          - type: "null"
      required:
      - personaId
    ProjectSelection:
      type: object
      properties:
        projectId:
          oneOf:
          - type: string
            format: uuid
          - type: "null"
    PersonaSelection:
      type: object
      properties:
        personaId:
          type: string
          format: uuid
      required:
      - personaId
...
    BranchSelection:
      type: object
      properties:
        messageId:
          type: string
          format: uuid
        expectedChildId:
          oneOf:
          - type: string
            format: uuid
          - type: "null"
      required:
      - messageId
...
    ChatBranch:
      type: object
      properties:
        id:
          type: string
          format: uuid
        parentMessageId:
          oneOf:
          - type: string
            format: uuid
          - type: "null"
        latestChildMessageId:
          oneOf:
          - type: string
            format: uuid
          - type: "null"
...
    ApiProblem:
      type: object
      additionalProperties: false
      properties:
        type:
          type: string
          format: uri
          description: Stable problem type for capability failures; omitted means
            RFC 9457 `about:blank`.
        title:
          type: string
          description: Short human-readable problem category.
        status:
          type: integer
          format: int32
          description: HTTP status code.
        detail:
          type: string
          description: Safe human-readable fallback detail.
        instance:
          type: string
          format: uri-reference
          description: Request path that produced the problem.
        code:
          type: string
          description: Stable capability-prefixed code; present only for expected
            capability failures.
          pattern: "^[A-Z][A-Z0-9_]*$"
        errors:
          type: array
          description: Field validation failures with safe allowlisted parameters;
            no rejected values.
          items:
            type: object
            additionalProperties: false
            properties:
              field:
                type: string
              message:
                type: string
                description: Compatible fallback message; clients localize by code.
              code:
                type: string
                enum:
                - REQUIRED
                - INVALID
                - EMAIL
                - SIZE
                - MIN
                - MAX
              params:
                type: object
                additionalProperties: false
                properties:
                  min:
                    type: number
                  max:
                    type: number
            required:
            - code
            - field
            - message
            - params
      required:
      - detail
      - instance
      - status
      - title
  securitySchemes:
    browserSession:
      type: apiKey
      description: HttpOnly JDBC-backed Spring Security session cookie.
      name: SESSION
      in: cookie
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT